requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.11.0
urllib3==2.2.3
web3==6.0.0
